feat: move checksum calculation into dbsync + improvements to recovery
