+++ src/shared_modules/dbsync/CMakeLists.txt
@@ -46,6 +46,7 @@ endif(APPLE)
 include_directories(${SRC_FOLDER}/external/sqlite/)
 include_directories(${SRC_FOLDER}/external/nlohmann/)
 include_directories(${SRC_FOLDER}/external/cJSON/)
+include_directories(${SRC_FOLDER}/external/openssl/include/)
 include_directories(${CMAKE_SOURCE_DIR}/include/)
 include_directories(${CMAKE_SOURCE_DIR}/src/)
 include_directories(${SHARED_MODULES}/utils/)
