+++ src/shared_modules/dbsync/tests/CMakeLists.txt
@@ -8,6 +8,7 @@ include_directories(${SRC_FOLDER}/external/googletest/googletest/include/)
 include_directories(${SRC_FOLDER}/external/googletest/googlemock/include/)
 
 link_directories(${SRC_FOLDER}/external/googletest/lib/)
+link_directories(${SRC_FOLDER}/external/openssl/)
 
 if(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   string(APPEND CMAKE_EXE_LINKER_FLAGS " -fprofile-arcs ")
