+++ src/win32/ossec.conf
@@ -179,6 +179,7 @@
       <enabled>yes</enabled>
       <interval>5m</interval>
       <max_eps>10</max_eps>
+      <integrity_interval>24h</integrity_interval>
     </synchronization>
   </wodle>
